move PomDependencies enum to tycho.core.shared
