+++ tycho-core/src/main/java/org/eclipse/tycho/core/TargetPlatformConfiguration.java
@@ -25,6 +25,7 @@ import org.eclipse.tycho.ArtifactKey;
 import org.eclipse.tycho.artifacts.TargetPlatformFilter;
 import org.eclipse.tycho.core.resolver.shared.IncludeSourceMode;
 import org.eclipse.tycho.core.resolver.shared.OptionalResolutionAction;
+import org.eclipse.tycho.core.resolver.shared.PomDependencies;
 import org.eclipse.tycho.core.shared.TargetEnvironment;
 import org.eclipse.tycho.p2.target.facade.TargetDefinitionFile;
 
@@ -32,23 +33,6 @@ public class TargetPlatformConfiguration implements DependencyResolverConfigurat
 
     public enum BREEHeaderSelectionPolicy {
         first, minimal
-    }
-
-    public enum PomDependencies {
-        /**
-         * pom dependencies are ignored
-         */
-        ignore,
-        /**
-         * pom dependencies are considered if the are already valid osgi artifacts. p2 metadata may
-         * be generated if missing
-         */
-        consider,
-        /**
-         * pom dependencies are used and wrapped into OSGi bundles if necessary. p2 metadata may be
-         * generated if missing.
-         */
-        wrapAsBundle;
     }
 
     private String resolver;
